# contracts/openapi/professor.yaml
openapi: 3.1.0
info:
  title: eduanima-professor API
  version: 1.0.0
  description: |
    Professor API (HTTP/JSON + SSE)

    ## 通信スタック
    - Frontend ↔ Professor: HTTP/JSON + SSE（永続）
    - Professor ↔ Librarian: gRPC双方向ストリーミング（Phase 1から）

    ## Phase 1: ローカル開発
    - 全リクエストで user_id='00000000-0000-0000-0000-000000000001' を固定使用
    - `X-Dev-User: dev-user` ヘッダーによる固定ユーザー識別
    - Phase 2でSSO認証実装時に /v1/auth/dev-login を削除し、Authorization: Bearer <token> に移行

    ## エンドポイント設計原則
    - リソースは名詞・複数形（subjects / materials / chats）
    - 階層ネスト: subjects 配下に materials・chats を配置
    - SSE: POST /v1/subjects/{id}/chats → Content-Type: text/event-stream で直接ストリーム
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/subjects/{subject_id}/files:
    post:
      tags: [OpenAI-Compat]
      summary: OpenAI Files 互換アップロード
      description: |
        OpenAI Files API 互換のファイルアップロードエンドポイント。
        purpose フィールドは受け取るが professor では無視される。
        アップロード後は GET /files/{file_id} で status が "processed" になるまでポーリングする。

        pageBench 用途:
          AGENT_FILE_STATUS_READY=processed に設定してポーリング。
      parameters:
        - $ref: '#/components/parameters/SubjectId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: アップロードするファイル
                purpose:
                  type: string
                  default: assistants
                  description: OpenAI 互換フィールド（professor では無視される）
      responses:
        '200':
          description: アップロード成功（OpenAI File オブジェクト）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/subjects/{subject_id}/files/{file_id}:
    get:
      tags: [OpenAI-Compat]
      summary: OpenAI File オブジェクト取得（ステータス確認）
      description: |
        OpenAI Files API 互換のファイルステータス確認エンドポイント。
        status が "processed" になれば chat/completions で使用可能。

        file_id は以下の形式を受け付ける:
          - OpenAI 互換形式: "file-550e8400e29b41d4a716446655440000"
          - 標準 UUID 形式: "550e8400-e29b-41d4-a716-446655440000"
      parameters:
        - $ref: '#/components/parameters/SubjectId'
        - name: file_id
          in: path
          required: true
          description: |
            File ID（OpenAI 形式 "file-{hex32}" または通常の UUID）
          schema:
            type: string
            example: "file-550e8400e29b41d4a716446655440000"
      responses:
        '200':
          description: File オブジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────
  # Health
  # ─────────────────────────────────────────
...
components:
  parameters:
    SubjectId:
      name: subject_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    OpenAIFile:
      type: object
      description: |
        OpenAI File オブジェクト互換。
        ID は "file-{hex32}" 形式（例: "file-550e8400e29b41d4a716446655440000"）。
      required: [id, object, bytes, created, filename, purpose, status]
      properties:
        id:
          type: string
          description: OpenAI 互換 File ID（"file-" + UUID ハイフンなし）
          example: "file-550e8400e29b41d4a716446655440000"
        object:
          type: string
          enum: [file]
        bytes:
          type: integer
          format: int64
          description: ファイルサイズ（バイト）
        created:
          type: integer
          format: int64
          description: Unix タイムスタンプ
        filename:
          type: string
          example: "lecture01.pdf"
        purpose:
          type: string
          description: アップロード時の purpose（pass-through）
          example: "assistants"
        status:
          type: string
          description: |
            OpenAI 互換のファイル処理状態。
            "uploaded"  → pending / processing（まだ処理中）
            "processed" → ready（検索可能 = AGENT_FILE_STATUS_READY の値）
            "error"     → failed（処理失敗）
          enum: [uploaded, processed, error]
        status_details:
          type: string
          nullable: true
          description: status="error" 時のエラー詳細
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: アプリケーションエラーコード（docs/03_integration/ERROR_CODES.md 参照）
              example: "VALIDATION_FAILED"
            message:
              type: string
              description: ユーザー向けメッセージ（内部詳細・Secretsを含まない）
              example: "Validation failed"
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              format: uuid